Rewrite.
The whole thing doesn't work in this way and I don't want to have such
cut and paste and bad looking nasty piece of code as an result of
enerving long and successless debugging.  So I abandon now and rewrite
some parts to have a pleasant and nice-looking peace of code
afterwards.