Finished alpha tests, corrected a few things.
I did the tests with two different data sets and stress tests.  While
doing this, I found some errors:
- The regular expression extracting the names from the game files
  accepts also numbers now.
Okay, it was just one mistake.

And there is a thing I forgot to say in the former commit: I removed
this obvious senseless, superfluous, redundant, waste and
unnecessecary thing with the template in the extra file.  Now its a
Here-Document.