Finished documentation.
I haven't commited for a long time since I forgot it.  Shame on me!
But now the main work is done and I have to check spelling and some
other things.  I have also corrected some things in the code, but it
doesn't work as expected.---It seems not liking a short help.  That's
a pity.